requirements.txt
pillow
numpy
streamlit
plotly
